+++ src/utils/distance.js
@@ -1,0 +1,3 @@
+export default (u, v) => {
+  return Math.hypot(u.x - v.x, u.y - v.y);
+};