+++ src/utils/step.js
@@ -1,0 +1,3 @@
+export default (limit, value) => {
+  return Number(value > limit);
+};